+++ src/main/java/com/spring/gradle/defect/repository/StatusWorkFlowRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.spring.gradle.defect.entity.StatusWorkFlow;
 @Repository
-public interface StatusWorkFlowRepo extends JpaRepository<StatusWorkFlow, Integer>{
+public interface StatusWorkFlowRepo extends JpaRepository<StatusWorkFlow, Long>{
 
 }
